% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibledata.R
\docType{data}
\name{bible}
\alias{bible}
\title{Bible data.}
\format{
A data frame with 31103 rows and 8 variables:
\describe{
  \item{id}{unique verse identifier}
  \item{book}{name of the boook}
  \item{chapter}{chapter of the verse}
  \item{verse}{number of the verse}
  \item{testament}{either old testament or new testament}
  \item{genre}{the type of literature of the bible the verse can be found in}
  \item{asv}{the American Standard Version of the Bible}
  \item{kjv}{the King James Version of the Bible}
  ...
}
}
\source{
\url{http://bible.com/}
}
\usage{
bible
}
\description{
A dataset containing every verse of the Bible with the genre, testament, and verse for
ASV and KJV (default).
}
\keyword{datasets}
