% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibles.R
\name{bibles}
\alias{bibles}
\title{Get a list of available Bible versions}
\usage{
bibles(
  language = NULL,
  abbreviation = NULL,
  name = NULL,
  ids = NULL,
  details = FALSE,
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)
}
\arguments{
\item{language}{ISO 639-3 three digit language code used to filter results}

\item{abbreviation}{Bible abbreviation to search for}

\item{name}{Bible name to search for}

\item{ids}{Comma separated list of Bible ids to return}

\item{details}{Boolean to include full Bible details (e.g. copyright and promo info)}

\item{debug}{Used to help debug the query}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will pull a list of all available bible versions
}
