% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{books}
\alias{books}
\title{Get a list of available books}
\usage{
books(
  bibleid = NULL,
  chapters = FALSE,
  sections = FALSE,
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)
}
\arguments{
\item{bibleid}{Identifies the Bible version (required)}

\item{chapters}{Boolean indicating if an array of chapter summaries should be included in the results. Defaults to false.}

\item{sections}{Boolean indicating if an array of chapter summaries and an array of sections should be included in the results. Defaults to false.}

\item{debug}{Used to help debug the query}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will pull a list of all available books in a particular version
}
