% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter.R
\name{chapter}
\alias{chapter}
\title{Get chapter contents}
\usage{
chapter(
  bibleid = Sys.getenv("MAIN_BIBLEID"),
  chapterid = NULL,
  contenttype = "json",
  includenotes = FALSE,
  includetitles = TRUE,
  includechapternumbers = FALSE,
  includeversenumbers = TRUE,
  includeversespans = FALSE,
  parallels = NULL,
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)
}
\arguments{
\item{bibleid}{Id of Bible whose Chapters to fetch}

\item{chapterid}{Id of the Chapter to fetch}

\item{contenttype}{Content type to be returned in the content property. Supported values are html (default), json (beta), and text (beta)}

\item{includenotes}{Include footnotes in content}

\item{includetitles}{Include section titles in content}

\item{includechapternumbers}{Include chapter numbers in content}

\item{includeversenumbers}{Include verse numbers in content.}

\item{includeversespans}{Include spans that wrap verse numbers and verse text for bible content.}

\item{parallels}{Comma delimited list of bibleIds to include}

\item{debug}{Used to help debug the query}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will pull a chapter cotent
}
