% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chapters}
\alias{chapters}
\title{Get a list of available chapters}
\usage{
chapters(
  bibleid = NULL,
  bookid = NULL,
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)
}
\arguments{
\item{bibleid}{Id of Bible whose Chapters to fetch}

\item{bookid}{Id of the Book whose Chapters to fetch}

\item{debug}{Used to help debug the query}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will pull a list of all available chapters
}
