% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingplan.R
\name{dayverse}
\alias{dayverse}
\title{Verse of the day}
\usage{
dayverse(date = Sys.Date())
}
\arguments{
\item{date}{the date which sets the seed for the random generator}
}
\description{
This function generates the verse of the day
}
