% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_definition.R
\name{define}
\alias{define}
\title{Reference word definitions}
\usage{
define(word = "test", key = Sys.getenv("WEBSTER_KEY"))
}
\arguments{
\item{word}{The desired word you need to look up}

\item{key}{The Webster dictionary API key}
}
\value{
A list of definitions
}
\description{
This function will pull in the definitions of a word from the Webster Mariam Collegiate dictionsary
}
