% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{frq}
\alias{frq}
\title{word frequency table}
\usage{
frq(txt)
}
\arguments{
\item{txt}{is the passage you want to analyze}
}
\value{
df of words and frequencies
}
\description{
This function will enable a user to return a verse by using just a book, chapter, and verse .
}
