% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lkp}
\alias{lkp}
\title{enables a search function}
\usage{
lkp(bk = "Genesis", ch = 1, vs = NA, translation = "kjv")
}
\arguments{
\item{bk}{book of the bible, not case senstaive}

\item{ch}{the chapter of the book}

\item{vs}{if a specific verse or number of verses, if you want the entire chapter then leave it blank}

\item{translation}{choose from kjv (default) or asv}
}
\description{
This function will enable me to search a phrase or word and return
the number of verses along with a quick reference.
}
