% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage.R
\name{passage}
\alias{passage}
\title{Get passage contents}
\usage{
passage(
  passageid = "mrk.1",
  contenttype = "text",
  includenotes = FALSE,
  includetitles = TRUE,
  includechapternumbers = FALSE,
  includeversenumbers = TRUE,
  includeversespans = FALSE,
  parallels = NULL,
  useorgid = FALSE,
  returnstring = TRUE,
  bibleid = Sys.getenv("MAIN_BIBLEID"),
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)
}
\arguments{
\item{passageid}{Id of the Chapter to fetch}

\item{contenttype}{Content type to be returned in the content property. Supported values are html (default), json (beta), and text (beta)}

\item{includenotes}{Include footnotes in content}

\item{includetitles}{Include section titles in content}

\item{includechapternumbers}{Include chapter numbers in content}

\item{includeversenumbers}{Include verse numbers in content.}

\item{includeversespans}{Include spans that wrap verse numbers and verse text for bible content.}

\item{parallels}{Comma delimited list of bibleIds to include}

\item{useorgid}{Use the supplied id(s) to match the verseOrgId instead of the verseId}

\item{returnstring}{Do you want a table or a string to be returned}

\item{bibleid}{Id of Bible whose Chapters to fetch}

\item{debug}{Used to help debug the query}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will pull a range of versions in a passage format
}
