% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{enables a search function}
\usage{
search(
  searchphrase = NA,
  translation = "kjv",
  testament_filter = c("OT", "NT"),
  ignorecase = TRUE
)
}
\arguments{
\item{searchphrase}{the pharase you want to seach}

\item{translation}{either kjv or asv at this time}

\item{testament_filter}{chooce from either OT or NT or both (default)}

\item{ignorecase}{Setting this to TRUE (defualt) will ignore whether the letters are upper or lwoer case}
}
\description{
This function will enable me to search a phrase or word and return
the number of verses along with a quick reference.
}
