% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{sections}
\alias{sections}
\title{Get a list of available sections}
\usage{
sections(
  bibleid = Sys.getenv("MAIN_BIBLEID"),
  bookid = NULL,
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)
}
\arguments{
\item{bibleid}{Id of Bible whose Chapters to fetch}

\item{bookid}{Id of the Book whose Chapters to fetch}

\item{debug}{Used to help debug the query}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will pull a list of all available sections
}
