% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{verse}
\alias{verse}
\title{Get a verse}
\usage{
verse(
  bibleid = Sys.getenv("MAIN_BIBLEID"),
  verseid = NULL,
  contenttype = "text",
  includenotes = FALSE,
  includetitles = TRUE,
  includechapternumbers = FALSE,
  includeversenumbers = TRUE,
  includeversespans = FALSE,
  parallels = c("55212e3cf5d04d49-01", "3aefb10641485092-01"),
  useorgid = FALSE,
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)
}
\arguments{
\item{bibleid}{Id of Bible whose Chapters to fetch}

\item{verseid}{Id of the verse to fetch}

\item{contenttype}{Content type to be returned in the content property. Supported values are html (default), json (beta), and text (beta)}

\item{includenotes}{Include footnotes in content}

\item{includetitles}{Include section titles in content}

\item{includechapternumbers}{Include chapter numbers in content}

\item{includeversenumbers}{Include verse numbers in content.}

\item{includeversespans}{Include spans that wrap verse numbers and verse text for bible content.}

\item{parallels}{Comma delimited list of bibleIds to include}

\item{useorgid}{Use the supplied id(s) to match the verseOrgId instead of the verseId}

\item{debug}{Used to help debug the query}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will pull a verse
}
