% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verses.R
\name{verses}
\alias{verses}
\title{Get a list of available verses}
\usage{
verses(
  chapterid = NULL,
  bibleid = Sys.getenv("MAIN_BIBLEID"),
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)
}
\arguments{
\item{chapterid}{Id of the Book whose Verses to fetch}

\item{bibleid}{Id of Bible whose Verses to fetch}

\item{debug}{Used to help debug the query}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will pull a list of all available verses
}
