% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vslookup.R
\name{vslkp}
\alias{vslkp}
\title{Looks up a verse and returns only the verse}
\usage{
vslkp(bk = "Genesis", ch = 1, vs = NA, translation = "kjv")
}
\arguments{
\item{bk}{book of the bible, not case senstaive}

\item{ch}{the chapter of the book}

\item{vs}{if a specific verse or number of verses, if you want the entire chapter then leave it blank}

\item{translation}{choose from kjv (default) or asv}
}
\value{
verse text
}
\description{
This function will enable a user to return a verse by using just a book, chapter, and verse .
}
