% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingplan.R
\name{readingplan}
\alias{readingplan}
\title{Get today's reading plan}
\usage{
readingplan(date = Sys.Date() - 2, apikey = Sys.getenv("BIBLER_APIKEY"))
}
\arguments{
\item{date}{defaults to today's date}

\item{apikey}{API.bible api key. Sign up here https://scripture.api.bible/signup}
}
\description{
This will return the passage for today's reading plan
}
